Initial commit
This commit contains updates to the gitignore and README files as well
as code. This is the initial pass and has a functional MVP. Upon running
`menu.py`, a list of all games present in the ROM directory will be
placed into a sorted list and shown on the LCD as the menu. The up &
down buttons on the LCD board can scroll through this list and the
select button can be used to load it onto the target device.

ROM directory and IP address for the target are currently hardcoded.