Revamp of the UI
The commits greatly revamps the UI/UX for NaomiPy. The flow is now
always driven through the game list first, and then once a game is
selected the target device can be. It's unintuitive to allow setting a
target device in a way that's decoupled from being game-driven, due to
the nature of the project.

Also adds a few points to the README for upcoming tasks.