Adds more descriptive yaml file for games
This commit adds a more fleshed out yaml file for the games list and
places it in a config directory. This directory will later be used to
allow for user configuration of settings and directories. The new game
list includes fields such as:
* System (NAOMI or Atomiswave)
* Orientation (horizontal or vertical)
* Requires peripheral (boolean)

This will allow more expressive sorting by the UI.

This commit also removes the old game lookup map.

This also heavily refactors how games are handled internally. It used to
use a flat mapping of display name to filename, but now it uses the more
expressive game definitions provided by config/game_list.yaml.

TODO:
* Test functionality with games.py